pyproject.toml
[tool.poetry]
name = "origin_api"
version = "0.1.0"
description = ""
authors = ["jeremie_j <jeremie@dsflow.io>"]

[tool.poetry.dependencies]
python = "^3.8"
SQLAlchemy = "^1.4.27"
pydantic = "^1.8.2"
python-dotenv = "^0.19.2"
psycopg2 = "^2.9.2"
requests = "^2.26.0"
bs4 = "^0.0.1"
lxml = "^4.6.4"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
psycopg2-binary = "^2.9.2"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
